+++ modules/events/RawEvent.php
@@ -1,5 +1,5 @@
 <?php
-  class @@CLASSNAME@@ {
+  class __CLASSNAME__ {
     public $name = "RawEvent";
 
     public function preprocessEvent($name, $registrations, $connection, $data) {